move deprecated things into deprecated module
